+++ oos-20190601/ChangeLog.md
@@ -1,3 +1,9 @@
+2024-11-08 Version: 3.2.3
+- Update API ListGitRepositories: add param OrgId.
+- Update API ListGitRepositories: update response param.
+- Update API ListTemplates: update response param.
+
+
 2024-08-16 Version: 3.2.2
 - Update API DeleteExecutions: add param Force.
 - Update API GetApplicationGroup: update response param.